requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2025.6
dj-database-url==3.0.0
Django==5.2.3
django-crispy-forms==2.4
django-environ==0.10.0
django-jazzmin==3.0.1
gunicorn==21.2.0
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
python-dotenv==1.0.0
sqlparse==0.5.3
typing_extensions==4.14.0
whitenoise==6.6.0
